doc: update sample app doc to use sudo
Previously the sample apps needed to be run as root, but with a recent
update they could be run as a normal (acrn) user.  Update the sample app
to use sudo for running these sample app components in case the user
doesn't login as root.  Also, split running the background app and the
foreground app into two commands to eliminate some confusion.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>