Makefile: drop the dependency on hvdefconfig
In the top-level Makefile, the target "hypervisor" depends on "hvdefconfig"
because, at the time the latter is introduced, building without specifying
BOARD or SCENARIO is supported by using a pair of default board and
scenario XMLs.

This is no longer the case today, as BOARD and SCENARIO are now mandatory
command line options. As a result, that dependency is no longer necessary,
and this patch just removes that.

Tracked-On: #8259
Signed-off-by: Junjie Mao <junjie.mao@intel.com>